+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat(repositories): added vlc repository
 - feat(repositories): added rpm-fusion repository
+- feat(repositories): added google-chrome repository
 
 ## 0.0.29 - 2026-08-09